test: Rename module to match conventions
