+++ lib/classes/CLI.test.js
@@ -623,7 +623,7 @@ describe('CLI', () => {
   describe('Integration tests', function() {
     this.timeout(1000 * 60 * 10);
     const that = this;
-    const serverlessExec = require('../../test/serverless-binary');
+    const serverlessExec = require('../../test/serverlessBinary');
     const env = resolveAwsEnv();
 
     before(() => {
