+++ lib/classes/PluginManager.test.js
@@ -2046,7 +2046,7 @@ describe('PluginManager', () => {
 
     let serverlessInstance;
     let serviceDir;
-    const serverlessExec = require('../../test/serverless-binary');
+    const serverlessExec = require('../../test/serverlessBinary');
 
     beforeEach(() => {
       // eslint-disable-line prefer-arrow-callback
