+++ test/integrationBasic.test.js
@@ -11,7 +11,7 @@ const resolveAwsEnv = require('@serverless/test/resolve-aws-env');
 const awsRequest = require('@serverless/test/aws-request');
 const { getTmpDirPath } = require('./utils/fs');
 
-const serverlessExec = require('./serverless-binary');
+const serverlessExec = require('./serverlessBinary');
 
 describe('Service Lifecyle Integration Test', function() {
   this.timeout(1000 * 60 * 10); // Involves time-taking deploys
